Removed unused code and consolidated code files
Deleted all the files inside the 3.5 project and used linked files.
Added conditional compilation symbol NET35 for CloudWatchAppender3.5
project.
