+++ CloudWatchAppender/Services/InstanceMetaDataReader.cs
@@ -152,7 +152,11 @@ namespace CloudWatchAppender.Services
                                               if (!_cachedValues.ContainsKey(key))
                                                   return "error_" + key;
 
+#if NET35
+                                              Debug.WriteLine(String.Format("Got {0}: {1}", key, _cachedValues[key]));
+#else
                                               Debug.WriteLine("Got {0}: {1}", key, _cachedValues[key]);
+#endif
                                               _pendingTasks.Remove(key);
                                               return _cachedValues[key];
                                           })
@@ -161,7 +165,11 @@ namespace CloudWatchAppender.Services
 
                 }
 
+#if NET35
+                Debug.WriteLine(String.Format("Returning cached {0}: {1}", key, _cachedValues[key]));
+#else
                 Debug.WriteLine("Returning cached {0}: {1}", key, _cachedValues[key]);
+#endif
                 return _cachedValues[key];
             }
             catch (WebException)
