+++ CloudWatchAppender/Services/ServiceTasks.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Concurrent;
 using System.Linq;
 using System.Threading.Tasks;
-using Amazon.Runtime;
 
 namespace CloudWatchAppender.Services
 {
