--- CloudWatchAppender3.5/TypeConverters/StandardUnitConverter.cs
@@ -1,20 +1,0 @@
-using System;
-using Amazon;
-using Amazon.CloudWatch;
-using log4net.Util.TypeConverters;
-
-namespace CloudWatchAppender.TypeConverters
-{
-    public class StandardUnitConverter :  IConvertFrom
-    {
-        public bool CanConvertFrom(Type sourceType)
-        {
-            return sourceType == typeof(string);
-        }
-
-        public object ConvertFrom(object source)
-        {
-            return StandardUnit.FindValue(source as string);
-        }
-    }
-}
